Rename PlanningPage component: update import path and file name for clarity
